refactor: 상품 ID 타입을 Int에서 Long으로 변경하고 데이터 바인딩을 사용하여 삭제 처리

+++ app/src/main/java/woowacourse/shopping/data/DatabaseCartRepository.kt
@@ -15,5 +15,5 @@ class DatabaseCartRepository
 
         override suspend fun getAllCartProducts(): List<Product> = cartDao.getAll().map { it.toDomain() }
 
-        override suspend fun deleteCartProduct(id: Int) = cartDao.delete(id.toLong())
+        override suspend fun deleteCartProduct(id: Long) = cartDao.delete(id)
     }
